mtk: mac80211: Add exported function for SoftMAC driver to get QoS map
Add exported function for SoftMAC driver to get QoS map.
1. Because the mapping from IP DSCP to IEEE 802.11 user priority may be customized.
Therefore, driver needs to pass the mapping to HW, so that the QoS type of traffic can be mapped in a consistent manner for both SW and HW paths.
2. due to this change(https://github.com/torvalds/linux/commit/6fdb8b8781d59796324efa25909f3e2112833f01) in backport 6.10.
we need to add a default QoS map. when hostapd config didn't set QoS map, we need pass default QoS map to HW.

Signed-off-by: Rex Lu <rex.lu@mediatek.com>

1. Remove exported function for SoftMAC driver to get QoS map.
Instead, add callback function for mac80211 to set QoS map in HW, which is more intuitive.
2. Fix inconsistent QoS mapping between AP and AP_VLAN IFs.
Specifically, when WDS AP IF is connected by a WDS STA, the QoS map of the AP_VLAN VIF is NULL.
So the QoS types of packets to the WDS STA will be determined using the default mapping rule.
However, SoftMAC driver uses the QoS map of the AP VIF, which may already be set.
Therefore, it is possible that the QoS mappings of SW and HW are inconsistent.
Thus, sync QoS map of AP VIF to that of AP_VLAN VIF.

Signed-off-by: Rex Lu <rex.lu@mediatek.com>

Refactor drv_set_qos_map function.
1. use dscp exception to instead of dscp range.
for example: if dscp value is 15. Original way will translte to tid 0.
but mac80211 will translate to tid 1.

Signed-off-by: Rex Lu <rex.lu@mediatek.com>
Signed-off-by: Benjamin Lin <benjamin-jw.lin@mediatek.com>

Do not copy Qos map from AP to AP_VLAN IF, because AP_VLAN's map will not be freed when IF is removed, causing memory leak.
Instead, assign pointer to AP's Qos map to AP_VLAN IF.
And add support for dynamic Qos-map setting for AP_VLAN IF.

Check if ieee80211_sub_if_data is ready before its usage to prevent later kernel warning.

Signed-off-by: Benjamin Lin <benjamin-jw.lin@mediatek.com>