mtk: mt76: mt7996: add external eeprom support
Add external eeprom support
For mt7992 and mt7990, efuse mode is not supported due to the lack of
space in efuse.
So, an additional external eeprom is added for user to store their
golden eeprom.

Note that the FW currently has some issues with writing to the ext
eeprom, so the write back function of ext eeprom is not yet linked
to any command.
A write back command will be added once the FW fixes the issue.

Add ext eeprom write callback

Signed-off-by: StanleyYP Wang <StanleyYP.Wang@mediatek.com>