mtk: mt76: mt7996: set eht max ampdu length
Set eht max ampdu length to 1. Without this patch, peer station cannot
get the correct eht capability and the 2G throughput cannot meet peak.

In the past, firmware does not check this field and set this field to 1
directly. However, it would lead to IOT issue when station cannot suppot
eht max ampdu length = 1. So firmware remove this issue and this field
need to be set by driver.

Signed-off-by: Peter Chiu <chui-hao.chiu@mediatek.com>