mtk: mt76: mt7996: config rxfilter to drop other unicast by default
Set MT_WF_RFCR_DROP_OTHER_UC by default and disable this flag in
mt7996_set_monitor only if monitor mode is enabled.

Without this patch, the MT_WF_RFCR_DROP_OTHER_UC would not be set so the
driver would receive lots of packets that is not sent to me.

Signed-off-by: Peter Chiu <chui-hao.chiu@mediatek.com>