mtk: mt76: mt7996: rework testmode dump stats
Rework testmode dump stats for dump order
Originally, the dump order is determined by the order of MT76 testmode
attribute enums, and the order of the attribute enum cannot be exchanged.
If a new attribute which is related to the previous attribute, the dump
info of these two attributes will be separated.
Therefore, use nla_for_each_nested instead of for loop to dump info, so
that the dump order will be determined by the order of nla_put
in driver.

Additionally, adjust the dump stats flow to adjust the order of dump
info.

Signed-off-by: StanleyYP Wang <StanleyYP.Wang@mediatek.com>